add custom zone name to obs detail view header
