+++ components/observations/ObservationDetailView.tsx
@@ -16,6 +16,7 @@ import {NACIcon} from 'components/icons/nac-icons';
 import {matchesZone} from 'components/observations/ObservationsFilterForm';
 import {AllCapsSm, AllCapsSmBlack, Body, BodyBlack, BodySemibold, bodySize} from 'components/text';
 import {HTML} from 'components/text/HTML';
+import {useMergedMapLayer} from 'hooks/useAlternateObservationZones';
 import {useAvalancheCenterCapabilities} from 'hooks/useAvalancheCenterCapabilities';
 import {useMapLayer} from 'hooks/useMapLayer';
 import {useNACObservation} from 'hooks/useNACObservation';
@@ -46,7 +47,7 @@ import {
   FormatSnowAvailableForTransport,
   FormatWindLoading,
   InstabilityDistribution,
-  MapLayer,
+  MergedMapLayer,
   Observation,
   SnowAvailableForTransport,
   WindLoading,
@@ -80,12 +81,13 @@ export const ObservationDetailView: React.FunctionComponent<{
   const mapLayer = mapResult.data;
   const capabilitiesResult = useAvalancheCenterCapabilities();
   const capabilities = capabilitiesResult.data;
+  const mergedMapLayer = useMergedMapLayer(observation?.center_id?.toUpperCase() as AvalancheCenterID, mapLayer);
 
-  if (incompleteQueryState(observationResult, mapResult, capabilitiesResult) || !observation || !mapLayer || !capabilities || !capabilities) {
+  if (incompleteQueryState(observationResult, mapResult, capabilitiesResult) || !observation || !mergedMapLayer || !capabilities || !capabilities) {
     return <QueryState results={[observationResult, mapResult, capabilitiesResult]} />;
   }
 
-  return <ObservationCard observation={observation} mapLayer={mapLayer} capabilities={capabilities} />;
+  return <ObservationCard observation={observation} mapLayer={mergedMapLayer} capabilities={capabilities} />;
 };
 
 const dataTableFlex = [1, 1];
@@ -201,7 +203,7 @@ export const withUnits = (value: string | number | null | undefined, units: stri
 
 export const ObservationCard: React.FunctionComponent<{
   observation: Observation;
-  mapLayer: MapLayer;
+  mapLayer: MergedMapLayer;
   capabilities: AllAvalancheCenterCapabilities;
 }> = ({observation, mapLayer, capabilities}) => {
   const navigation = useNavigation<ObservationsStackNavigationProps>();
